requirements.txt
datasets==2.12.0
h5py==3.9.0
hf-argparser @ git+https://github.com/clane9/hf-argparser@v4.29.0
ipython==8.10.0
jupyter==1.0.0
matplotlib==3.7.0
numpy==1.24.2
opencv-python-headless==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
pre-commit==3.3.3
pycocotools==2.0.6
pycortex==1.2.6
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
shapely==2.0.1
timm==0.9.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
wandb==0.15.2
